+++ ValheimExportHelper/LoggingTrait.cs
@@ -1,22 +1,24 @@
-﻿using AssetRipper.Core.Logging;
-
-namespace ValheimExportHelper
+﻿namespace ValheimExportHelper
 {
   public abstract class LoggingTrait
   {
     public void LogInfo(string text)
     {
-      Logger.Info(LogCategory.Plugin, $"[{GetType().FullName}] {text}");
+      Console.WriteLine($"[{GetType().FullName}] {text}");
     }
 
     public void LogWarn(string text)
     {
-      Logger.Warning(LogCategory.Plugin, $"[{GetType().FullName}] {text}");
+      Console.ForegroundColor = ConsoleColor.Yellow;
+      Console.WriteLine($"[WARN] [{GetType().FullName}] {text}");
+      Console.ResetColor();
     }
 
     public void LogError(string text)
     {
-      Logger.Error(LogCategory.Plugin, $"[{GetType().FullName}] {text}");
+      Console.ForegroundColor = ConsoleColor.Red;
+      Console.WriteLine($"[ERROR] [{GetType().FullName}] {text}");
+      Console.ResetColor();
     }
   }
 }