+++ ValheimExportHelper/Ripper.cs
@@ -1,0 +1,8 @@
+﻿namespace ValheimExportHelper
+{
+  class Ripper
+  {
+    public string ExportRootPath;
+    public string GameDataPath;
+  }
+}